WIP: device: drop NM_UNMANAGED_EXTERNAL_DOWN
We used the unmanage flag "external-down" for software devices that
were created outside of NetworkManager. The "external-down" flag kept
the device as unmanaged, until the interface was up and had an IP address.
That is wrong to do.

Instead, replace the flag by NM_UNMANAGED_BY_DEFAULT, which marks all
software devices that were created outside of NetworkManager as unmanaged.
This is regardless of their IP configuration. If a software device was not
created by NetworkManager it is fair to assume that the user wants to manage
it himself and doesn't want NetworkManager to interfere.

Note that the unmanged flags "user-udev" and "user-explicit" have higher
priority then the "by-default" flag. This way it will be still possible to
manage devices that would be otherwise unmanaged.

This breaks behavior in many ways, which will be only fixed by later commits.