core/trivial: rename nm-generated-assumed flag to garbage-collect
The concept of assumed-connection will eventually go away (as it existed
until now). What is still useful is the property of a connection to get
deleted once the connection disconnects.

That is, a settings connection lingers around as long as it's active,
but once it deactivates it gets automatically deleted. As such, it's
a more volatile concept of an in-memory connection.

This is related to bug rh#1401515