core: track external activations types in the active-connection
We need a distinction between external activations and assuming
connections. The former shall have the meaning of devices that are
*not* managed by NetworkManager, the latter are configurations that
are gracefully taken over after restart (but fully managed).

Express that in the activation-type of the active connection.

Also, no longer use the settings NM_SETTINGS_CONNECTION_FLAGS_GARBAGE_COLLECT
flag to determine whether an assumed connection is "external". These
concepts are entirely orthogonal (although in pratice, external
activations are in-memory and flagged for garbage collection.

The change to match_connection_filter() to also consider connections
that are marked for garbagecollect will later make more sense,
where we only do nm_utils_match_connection() for assumed connection
but skip it for external activation-type.