core/dispatcher: refactor device dispatcher calls
Currently, we determine NMD_CONNECTION_PROPS_EXTERNAL based
on the settings connection. That is wrong, because whether
a connection is assumed/external, really is a property of
the active-connection. So, in the future, we would need yet
another argument to nm_dispatcher_call(), as the external
flag is part of the NMActRequest.

Instead, drop the settings-connection and applied-connection
arguments. They are redundant.

Also, rename nm_dispatcher_call() to nm_dispatcher_call_device(),
it this is not a generic dispatcher call, but it is particularly
for a device.

Also, rename nm_dispatcher_call_sync() to nm_dispatcher_call_device_sync().