dns: introduce routing domains
Similarly to what systemd-resolved does, introduce the concept of
"routing-only" domain, which is a domain in the search list that is
not used to complete unqualified host names, but only to decide the
interface over which a query must be forwarded. Routing domains are
those starting with a tilde ('~') before the actual domain name.

Domains without the initial tilde are used both for completing
unqualified names and for routing decision.