dns: support wildcard routing domain
The special wildcard routing domain '~.' can be used to prefer the
interface(s) that specify it when a query must be sent for a domain
that does not match any specific domain.

When no wildcard domain is present, queries without a specific match
are sent to all interfaces, except VPNs that provide a search list.