setting-wireless: add wps property
This property will be used to decide if it makes sense to attempt a WPS
enrollment on connection activation.

In general it will make very little sense to set anything here manually,
but when we'll auto-complete a Wi-Fi connection to a specific BSS we'll
be able to set this to a sane value automatically. This also is the
reason we don't bother exposing the property in CLI or saving by the
setting plugins.